% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Concrete}
\alias{Concrete}
\title{Concrete - This is the strength of concrete daa set originally posted on UCI}
\format{
Concrete
A data frame with 1030 rows and 9 columns:
\describe{
   \item{Cement}{ quantitative -- kg in a m3 mixture -- Input Variable}
   \item{Blast_Furnace_Slag}{quantitative -- kg in a m3 mixture -- Input Variable}
   \item{Fly_Ash}{ quantitative  -- kg in a m3 mixture -- Input Variable}
   \item{Water}{quantitative  -- kg in a m3 mixture -- Input Variable}
   \item{Superplasticizer}{quantitative -- kg in a m3 mixture -- Input Variable}
   \item{Coarse_Aggregate}{quantitative -- kg in a m3 mixture -- Input Variable}
   \item{Fine_Aggregate}{quantitative  -- kg in a m3 mixture -- Input Variable}
   \item{Age}{Day (1~365) -- Input Variable}
   \item{Strength}{quantitative -- MPa -- Output Variable}
}
}
\source{
https://archive.ics.uci.edu/dataset/165/concrete+compressive+strength
}
\usage{
Concrete
}
\description{
Concrete is the most important material in civil engineering. The concrete compressive strength is a highly nonlinear function of age and ingredients.
}
\keyword{datasets}
