% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumericEnsembles.R
\name{ExportNumericResults}
\alias{ExportNumericResults}
\title{Export Numeric Pipeline Outcomes and Graphs}
\usage{
ExportNumericResults(pipeline_object = NULL, export_directory = NULL)
}
\arguments{
\item{pipeline_object}{An object generated by the \code{Numeric()} engine.}

\item{export_directory}{Character string path targeting file output folders.}
}
\value{
Invisible character string tracking file save directory destinations.
}
\description{
Export Numeric Pipeline Outcomes and Graphs
}
