% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Insurance}
\alias{Insurance}
\title{Insurance - The data is from UCI}
\format{
Insurance
A data frame with 1338 rows and 7 columns
Credit to Bob Wakefield
\describe{
   \item{Age}{The age of the customer. (Integer)}
   \item{Children}{The number of children the customer has. (Integer)}
   \item{Smoker}{Whether or not the customer is a smoker. (Boolean)}
   \item{Region}{The region the customer lives in. (String)}
   \item{Charges}{The insurance charges for the customer. (Float)}
}
}
\source{
https://www.kaggle.com/datasets/thedevastator/prediction-of-insurance-charges-using-age-gender
}
\usage{
Insurance
}
\description{
This dataset contains detailed information about insurance customers, including their age, sex, body mass index (BMI), number of children, smoking status and region. Having access to such valuable insights allows analysts to get a better view into customer behaviour and the factors that contribute to their insurance charges.
}
\keyword{datasets}
