% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumericEnsembles.R
\name{NumericEnsemblesFastConfig}
\alias{NumericEnsemblesFastConfig}
\title{Fast-Execution Configuration Matrix for Regression Testing}
\usage{
NumericEnsemblesFastConfig()
}
\value{
A fast-track parameter subset list optimized for rapid package verification checks.
}
\description{
Fast-Execution Configuration Matrix for Regression Testing
}
