% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumericEnsembles.R
\name{RenderExecutiveReport}
\alias{RenderExecutiveReport}
\title{Render Automated Quarto Numeric Executive Report}
\usage{
RenderExecutiveReport(pipeline_object = NULL, output_directory = getwd())
}
\arguments{
\item{pipeline_object}{An object generated by the \code{Numeric()} engine.}

\item{output_directory}{Character string specifying where to save the report files.}
}
\description{
Render Automated Quarto Numeric Executive Report
}
