% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumericEnsembles.R
\name{ExportNumericResults}
\alias{ExportNumericResults}
\title{Export Tables and High-Resolution Visualization Panels}
\usage{
ExportNumericResults(pipeline_object = NULL, export_directory = NULL)
}
\arguments{
\item{pipeline_object}{A trained \code{numeric_pipeline} asset.}

\item{export_directory}{Character path detailing target output folder destination.}
}
\value{
Invisible string path location where files were compiled.
}
\description{
Export Tables and High-Resolution Visualization Panels
}
