% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumericEnsembles.R
\name{LaunchNumericApp}
\alias{LaunchNumericApp}
\title{Launch Interactive Shiny Application Portal}
\usage{
LaunchNumericApp()
}
\value{
Activated browser session running the dashboard interface.
}
\description{
Launch Interactive Shiny Application Portal
}
