% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumericEnsembles.R, R/NumericEnsemblesBACKUP.R
\name{LaunchNumericApp}
\alias{LaunchNumericApp}
\title{Launch Interactive NumericEnsembles Web Interface App}
\usage{
LaunchNumericApp()

LaunchNumericApp()
}
\description{
Fires up a local instance of the interactive OLS vs. GLM tuning,
diagnostic, and residual validation web dashboard system.

Fires up a local instance of the interactive OLS vs. GLM tuning,
diagnostic, and residual validation web dashboard system.
}
