% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumericEnsembles.R, R/NumericEnsemblesBACKUP.R
\name{load_pipeline}
\alias{load_pipeline}
\title{Load Serialized Numeric Pipeline Environment from Disk File}
\usage{
load_pipeline(file_path)

load_pipeline(file_path)
}
\arguments{
\item{file_path}{Character string tracking destination directory file path.}
}
\value{
A re-hydrated numeric_pipeline object.

A re-hydrated numeric_pipeline object.
}
\description{
Load Serialized Numeric Pipeline Environment from Disk File

Load Serialized Numeric Pipeline Environment from Disk File
}
