% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumericEnsembles.R
\name{Numeric}
\alias{Numeric}
\title{Numeric—function to automatically build 23 individual models and 17 ensembles then return the results to the user}
\usage{
Numeric(
  data,
  colnum,
  numresamples,
  how_to_handle_strings = c(0("none"), 1("factor levels")),
  do_you_have_new_data = c("Y", "N"),
  save_all_trained_models = c("Y", "N"),
  remove_ensemble_correlations_greater_than,
  use_parallel = c("Y", "N"),
  train_amount,
  test_amount,
  validation_amount
)
}
\arguments{
\item{data}{data can be a CSV file or within an R package, such as MASS::Boston}

\item{colnum}{a column number in your data}

\item{numresamples}{the number of resamples}

\item{how_to_handle_strings}{0: No strings, 1: Factor values, 2: One-hot encoding, 3: One-hot encoding AND jitter}

\item{do_you_have_new_data}{"Y" or "N". If "Y", then you will be asked for the new data}

\item{save_all_trained_models}{"Y" or "N". If "Y", then places all the trained models in the Environment}

\item{remove_ensemble_correlations_greater_than}{Enter a number to remove correlations in the ensembles}

\item{use_parallel}{"Y" or "N" for parallel processing}

\item{train_amount}{set the amount for the training data}

\item{test_amount}{set the amount for the testing data}

\item{validation_amount}{Set the amount for the validation data}
}
\value{
a real number
}
\description{
Numeric—function to automatically build 23 individual models and 17 ensembles then return the results to the user
}
\examples{
# Note that examples take about one minute each to run to completion
Numeric(data = Boston_housing,
  colnum = 9,
  numresamples = 2,
  how_to_handle_strings = 0,
  do_you_have_new_data = "N",
  save_all_trained_models = "N",
  remove_ensemble_correlations_greater_than = 1.00,
  train_amount = 0.60,
  test_amount = 0.20,
  validation_amount = 0.20)

 Numeric(data = Concrete,
  colnum = 9,
  numresamples = 2,
  how_to_handle_strings = 1,
  do_you_have_new_data = "N",
  save_all_trained_models = "N",
  remove_ensemble_correlations_greater_than = 1.00,
  train_amount = 0.60,
  test_amount = 0.20,
  validation_amount = 0.20
  )

 Numeric(data = Insurance,
  colnum = 7,
  numresamples = 2,
  how_to_handle_strings = 1,
  do_you_have_new_data = "N",
  save_all_trained_models = "N",
  remove_ensemble_correlations_greater_than = 1.00,
  train_amount = 0.60,
  test_amount = 0.20,
  validation_amount = 0.20)

}
