% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumericEnsembles.R
\name{NumericEnsemblesDemo}
\alias{NumericEnsemblesDemo}
\title{Run Fast Validation Regression Demo Suite}
\usage{
NumericEnsemblesDemo()
}
\value{
An invisible trained \code{numeric_pipeline} object run over simulated macroeconomic metrics.
}
\description{
Run Fast Validation Regression Demo Suite
}
