% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumericEnsembles.R
\name{NumericEnsemblesFastConfig}
\alias{NumericEnsemblesFastConfig}
\title{Fast-Execution Configuration Matrix}
\usage{
NumericEnsemblesFastConfig()
}
\value{
A structured list containing optimized, fast-track hyperparameter settings.
}
\description{
Fast-Execution Configuration Matrix
}
