% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumericEnsembles.R, R/NumericEnsemblesBACKUP.R
\name{predict.numeric_pipeline}
\alias{predict.numeric_pipeline}
\title{Predict with Numeric Pipeline Framework}
\usage{
\method{predict}{numeric_pipeline}(object, newdata, model_name = "best", ...)

\method{predict}{numeric_pipeline}(object, newdata, model_name = "best", ...)
}
\arguments{
\item{object}{A trained \code{numeric_pipeline} object.}

\item{newdata}{A data.frame containing new data configurations to score.}

\item{model_name}{Character string specifying the target model from the leaderboard to score. Default = "best".}

\item{...}{Additional arguments passed to underlying predict methods.}
}
\value{
A numeric vector of predictions.

A numeric vector of predictions.
}
\description{
Generates predictions across new datasets matching models evaluated in the pipeline.

Generates predictions across new datasets matching models evaluated in the pipeline.
}
