% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumericEnsembles.R
\name{predict_production}
\alias{predict_production}
\title{Generate Executive Production Projections with Row-Level 95 percent Confidence Intervals}
\usage{
predict_production(object, newdata)
}
\arguments{
\item{object}{A trained `numeric_pipeline` object.}

\item{newdata}{A data.frame containing new data configurations to score.}
}
\value{
A data.frame structured with prediction vectors and interval metrics for the top 3 models.
}
\description{
Generate Executive Production Projections with Row-Level 95 percent Confidence Intervals
}
