% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumericEnsembles.R
\name{print.numeric_pipeline}
\alias{print.numeric_pipeline}
\title{Print Numeric Pipeline Summary Profiles Report}
\usage{
\method{print}{numeric_pipeline}(x, ...)
}
\arguments{
\item{x}{A numeric_pipeline object generated by the Numeric() engine.}

\item{...}{Additional arguments passed to underlying print methods.}
}
\description{
Print Numeric Pipeline Summary Profiles Report
}
