% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumericEnsembles.R
\name{RenderExecutiveReport}
\alias{RenderExecutiveReport}
\title{Compile and Render Automated Executive Summary Report}
\usage{
RenderExecutiveReport(pipeline_object, output_directory = getwd())
}
\arguments{
\item{pipeline_object}{A trained \code{numeric_pipeline} asset.}

\item{output_directory}{Character path mapping where summary files are rendered.}
}
\value{
Invisible NULL.
}
\description{
Compile and Render Automated Executive Summary Report
}
