% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumericEnsembles.R, R/NumericEnsemblesBACKUP.R
\name{save_pipeline}
\alias{save_pipeline}
\title{Save Serialized Numeric Pipeline Environment to Disk File}
\usage{
save_pipeline(object, file_path)

save_pipeline(object, file_path)
}
\arguments{
\item{object}{A numeric_pipeline object.}

\item{file_path}{Character string tracking destination directory file path.}
}
\description{
Save Serialized Numeric Pipeline Environment to Disk File

Save Serialized Numeric Pipeline Environment to Disk File
}
